+++ xbmc/platform/darwin/DarwinUtils.h
@@ -42,8 +42,8 @@ public:
   static float       GetIOSVersion(void);
   static const char *GetIOSVersionString(void);
   static const char *GetOSXVersionString(void);
-  static int         GetFrameworkPath(char* path, uint32_t *pathsize);
-  static int         GetExecutablePath(char* path, uint32_t *pathsize);
+  static int         GetFrameworkPath(char* path, size_t *pathsize);
+  static int         GetExecutablePath(char* path, size_t *pathsize);
   static const char *GetAppRootFolder(void);
   static bool        IsIosSandboxed(void);
   static bool        HasVideoToolboxDecoder(void);
